MOSAIKS_feature/requirements.txt
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.3
geopandas==1.0.1
contextily==1.6.2
folium==0.19.2
seaborn==0.13.2
geopy==2.4.1
scikit-learn==1.6.0
shapely==2.0.6
pillow==11.0.0
pytest==8.3.4